@@ -5,7 +5,7 @@ publish_to: 'none'
 dependencies:
   ethereum_codec:
     path: ../
-  dio: ^3.0.10
+  dio: ^4.0.6
 
 environment:
-  sdk: ">=1.19.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
