@@ -1,19 +1,17 @@
 name: ethereum_codec
-version: 0.3.0
+version: 2.0.0
 publish_to: none
 
 dependencies:
-  convert: ^2.1.1
-  pointycastle: ^1.0.1
-  typed_data: ^1.2.0
-  sprintf: ^5.0.0
-  ethereum_util:
-    git:
-      url: https://github.com/ethereumdart/ethereum_util
+  convert: ^3.1.1
+  pointycastle: ^3.6.2
+  typed_data: ^1.3.1
+  sprintf: ^7.0.0
+
   eth_abi_codec:
     git:
-     url: https://github.com/nbltrust/dart-eth-abi-codec
-     ref: '0.2.1'
+      url: https://github.com/nbltrust/dart-eth-abi-codec.git
+      ref: 2.0.1
 
 environment:
-  sdk: ">=1.19.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
