@@ -5,10 +5,10 @@ publish_to: 'none'
 dependencies:
   ethereum_codec:
     path: ../
-  decimal: ^0.3.0
-  dio: ^3.0.10
-  test: any
-  intl: ^0.16.0
+  decimal: ^2.3.0
+  dio: ^4.0.6
+  test: ^1.21.6
+  intl: ^0.17.0
 
 environment:
-  sdk: ">=1.19.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
