@@ -7,7 +7,7 @@ import 'package:ethereum_codec/ethereum_codec.dart';
 Future<dynamic> callInfura(String method, List<dynamic> args) async {
   var dio = new Dio();
   var response = await dio.post(
-    'https://mainnet.infura.io/v3/774b1e4252de48c3997d66ac5f5078d8',
+    'https://mainnet.infura.io/v3/12a9161db1dd4fce8290ab19ea9128d5',
     data: {
       'id': 1,
       'jsonrpc': '2.0',
@@ -25,15 +25,16 @@ Future<dynamic> callInfura(String method, List<dynamic> args) async {
 void initAbi() {
   var symbolFile = File('contract_abi/contract_symbols.json');
   var symbols = jsonDecode(symbolFile.readAsStringSync());
-  var abis = (symbols as List).map((s) => <String, dynamic>{
+  List<dynamic> contracts = symbols["contracts"];
+  var abis = contracts.map((s) => <String, dynamic>{
     'type': s['type'] as String,
     'abi': jsonDecode(
       File('contract_abi/abi/${s['type'].toUpperCase()}.json').readAsStringSync())
   }).toList();
 
-  var transFile = File('contract_abi/trx_trans.json');
-  var translators = jsonDecode(transFile.readAsStringSync());
-  initContractABIs(symbols, abis, translators: translators);
+  // var transFile = File('contract_abi/trx_trans.json');
+  // var translators = jsonDecode(transFile.readAsStringSync());
+  initContractABIs(contracts, abis,);
 
   ETHRpc.createInstance(callInfura);
 }
