@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:convert/convert.dart';
 import 'package:decimal/decimal.dart';
@@ -18,8 +19,11 @@ class ParsedResult {
 Future<ParsedResult> parseSwap(ContractCall callInfo, dynamic tx) async {
   final params = callInfo.callParams;
   final recipient = '0x' + params['to'].toString();
-
-  final amountInKey = params.keys.firstWhere((k) => k.startsWith('amountIn'), orElse: () => null);
+  print("---------------- $params");
+  final amountInKey = params.keys.cast<String?>().firstWhere((k) {
+    if(k == null) return false;
+   return k.startsWith('amountIn');
+  }, orElse: () => null);
   final amountIn = amountInKey != null ? Decimal.parse(params[amountInKey].toString()) : Decimal.parse(tx.value.toString());
 
   final amountOutKey = params.keys.firstWhere((k) => k.startsWith('amountOut'));
@@ -31,7 +35,7 @@ Future<ParsedResult> parseSwap(ContractCall callInfo, dynamic tx) async {
 
   final tokenIn = getContractConfigByAddress(tokenInAddr);
   final tokenOut = getContractConfigByAddress(tokenOutAddr);
-
+print("-------------tokenOut $tokenOut   tokenIn $tokenIn");
   if (tokenIn == null || tokenOut == null) {
     return ParsedResult(recipient, []);
   }
@@ -47,8 +51,8 @@ Future<ParsedResult> parseSwap(ContractCall callInfo, dynamic tx) async {
   }
 
   final args = [
-    [amountOutKey.endsWith('Min') ? '' : '≤', amountIn / Decimal.parse('1e' + tokenIn.params['decimal']), symbolIn],
-    [amountOutKey.endsWith('Min') ? '≥' : '', amountOut / Decimal.parse('1e' + tokenOut.params['decimal']), symbolOut],
+    [amountOutKey.endsWith('Min') ? '' : '≤', (amountIn / Decimal.parse('1e' + tokenIn.params?['decimal'])).toDecimal().toString(), symbolIn],
+    [amountOutKey.endsWith('Min') ? '≥' : '', (amountOut / Decimal.parse('1e' + tokenOut.params?['decimal'])).toDecimal().toString(), symbolOut],
   ];
 
   return ParsedResult(recipient, ['swap', args[0], args[1]]);
@@ -91,11 +95,11 @@ void main() async {
         int.parse('2d386', radix: 16),
         int.parse('11abfe1700', radix: 16),
         int.parse('536', radix: 16),
-        input: hex.decode(
-            "fb3bdb410000000000000000000000000000000000000000000000000522810a26e500000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000a31dbf0435af02f3b68ec7f985c9388e8ab1e47b000000000000000000000000000000000000000000000000000000006125eeba0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000114f1388fab456c4ba31b1850b244eedcd024136"));
+        input: Uint8List.fromList(hex.decode(
+            "fb3bdb410000000000000000000000000000000000000000000000000522810a26e500000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000a31dbf0435af02f3b68ec7f985c9388e8ab1e47b000000000000000000000000000000000000000000000000000000006125eeba0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000114f1388fab456c4ba31b1850b244eedcd024136")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -114,11 +118,11 @@ void main() async {
         int.parse('306bc', radix: 16),
         int.parse('f852e00a6', radix: 16),
         int.parse('2b3', radix: 16),
-        input: hex.decode(
-            "7ff36ab500000000000000000000000000000000000000000000000007aa585f0431278c000000000000000000000000000000000000000000000000000000000000008000000000000000000000000087c8b347838d9d21a14aeb29cf3437098a87a80c000000000000000000000000000000000000000000000000000000006125f5940000000000000000000000000000000000000000000000000000000000000002000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000098968f0747e0a261532cacc0be296375f5c08398"));
+        input: Uint8List.fromList(hex.decode(
+            "7ff36ab500000000000000000000000000000000000000000000000007aa585f0431278c000000000000000000000000000000000000000000000000000000000000008000000000000000000000000087c8b347838d9d21a14aeb29cf3437098a87a80c000000000000000000000000000000000000000000000000000000006125f5940000000000000000000000000000000000000000000000000000000000000002000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000098968f0747e0a261532cacc0be296375f5c08398")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -137,11 +141,11 @@ void main() async {
         int.parse('29e33', radix: 16),
         int.parse('10c0837980', radix: 16),
         int.parse('11', radix: 16),
-        input: hex.decode(
-            "18cbafe500000000000000000000000000000000000000000000001b1ae4d6e2ef50000000000000000000000000000000000000000000000000000005be0fa12b2ce6d400000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000bab53f3f51e67a1b3da487631c88757addc0296c0000000000000000000000000000000000000000000000000000000061260da80000000000000000000000000000000000000000000000000000000000000002000000000000000000000000a80f2c8f61c56546001f5fc2eb8d6e4e72c45d4c000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"));
+        input: Uint8List.fromList(hex.decode(
+            "18cbafe500000000000000000000000000000000000000000000001b1ae4d6e2ef50000000000000000000000000000000000000000000000000000005be0fa12b2ce6d400000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000bab53f3f51e67a1b3da487631c88757addc0296c0000000000000000000000000000000000000000000000000000000061260da80000000000000000000000000000000000000000000000000000000000000002000000000000000000000000a80f2c8f61c56546001f5fc2eb8d6e4e72c45d4c000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -160,11 +164,11 @@ void main() async {
         int.parse('3b8a6', radix: 16),
         int.parse('eab17b600', radix: 16),
         int.parse('2c', radix: 16),
-        input: hex.decode(
-            "38ed17390000000000000000000000000000000000000000000000182d7e4cfda0380000000000000000000000000000000000000000000000000000000000035378eefd00000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000f746db80998cf53483c0a8ce895f6c69438fe847000000000000000000000000000000000000000000000000000000006125f9ee00000000000000000000000000000000000000000000000000000000000000030000000000000000000000002e9d63788249371f1dfc918a52f8d799f4a38c94000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"));
+        input: Uint8List.fromList(hex.decode(
+            "38ed17390000000000000000000000000000000000000000000000182d7e4cfda0380000000000000000000000000000000000000000000000000000000000035378eefd00000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000f746db80998cf53483c0a8ce895f6c69438fe847000000000000000000000000000000000000000000000000000000006125f9ee00000000000000000000000000000000000000000000000000000000000000030000000000000000000000002e9d63788249371f1dfc918a52f8d799f4a38c94000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -183,11 +187,11 @@ void main() async {
         int.parse('2a084', radix: 16),
         int.parse('e0fd67c49', radix: 16),
         int.parse('a', radix: 16),
-        input: hex.decode(
-            "4a25d94a00000000000000000000000000000000000000000000000006f05b59d3b200000000000000000000000000000000000000000000000001ea51418f1a706afff100000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000af877d174d4c3a8de4f412a65dd089f94be37ed7000000000000000000000000000000000000000000000000000000006125f34b000000000000000000000000000000000000000000000000000000000000000200000000000000000000000006f3c323f0238c72bf35011071f2b5b7f43a054c000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"));
+        input: Uint8List.fromList(hex.decode(
+            "4a25d94a00000000000000000000000000000000000000000000000006f05b59d3b200000000000000000000000000000000000000000000000001ea51418f1a706afff100000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000af877d174d4c3a8de4f412a65dd089f94be37ed7000000000000000000000000000000000000000000000000000000006125f34b000000000000000000000000000000000000000000000000000000000000000200000000000000000000000006f3c323f0238c72bf35011071f2b5b7f43a054c000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -206,11 +210,11 @@ void main() async {
         int.parse('82554', radix: 16),
         int.parse('faddb72e8', radix: 16),
         int.parse('8e5', radix: 16),
-        input: hex.decode(
-            "8803dbee000000000000000000000000000000000000000000000007e50ccc6899b00000000000000000000000000000000000000000000000000000000000011e110e5f00000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ea1fcb26d0f438e331559e17a07f94d478ad945a000000000000000000000000000000000000000000000000000000006125fc270000000000000000000000000000000000000000000000000000000000000003000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000fca59cd816ab1ead66534d82bc21e7515ce441cf"));
+        input: Uint8List.fromList(hex.decode(
+            "8803dbee000000000000000000000000000000000000000000000007e50ccc6899b00000000000000000000000000000000000000000000000000000000000011e110e5f00000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ea1fcb26d0f438e331559e17a07f94d478ad945a000000000000000000000000000000000000000000000000000000006125fc270000000000000000000000000000000000000000000000000000000000000003000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000fca59cd816ab1ead66534d82bc21e7515ce441cf")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -229,11 +233,11 @@ void main() async {
         int.parse('6ddd0', radix: 16),
         int.parse('20fd78b5e1', radix: 16),
         int.parse('3215', radix: 16),
-        input: hex.decode(
-            "791ac9470000000000000000000000000000000000000000000000372579316ecf08ade6000000000000000000000000000000000000000000000000179a8d281873fe0000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ae7861c80d03826837a50b45aecf11ec677f6586000000000000000000000000000000000000000000000000000000006125b9780000000000000000000000000000000000000000000000000000000000000002000000000000000000000000ae12c5930881c53715b369cec7606b70d8eb229f000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"));
+        input: Uint8List.fromList(hex.decode(
+            "791ac9470000000000000000000000000000000000000000000000372579316ecf08ade6000000000000000000000000000000000000000000000000179a8d281873fe0000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ae7861c80d03826837a50b45aecf11ec677f6586000000000000000000000000000000000000000000000000000000006125b9780000000000000000000000000000000000000000000000000000000000000002000000000000000000000000ae12c5930881c53715b369cec7606b70d8eb229f000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -252,11 +256,11 @@ void main() async {
         int.parse('6ddd0', radix: 16),
         int.parse('13532f7e01', radix: 16),
         int.parse('3112', radix: 16),
-        input: hex.decode(
-            "5c11d795000000000000000000000000000000000000000000000d80073b37fc0d9e39e200000000000000000000000000000000000000000000000000000001677f5af500000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ae7861c80d03826837a50b45aecf11ec677f6586000000000000000000000000000000000000000000000000000000006125b3cc0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000a487bf43cf3b10dffc97a9a744cbb7036965d3b9000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7"));
+        input: Uint8List.fromList(hex.decode(
+            "5c11d795000000000000000000000000000000000000000000000d80073b37fc0d9e39e200000000000000000000000000000000000000000000000000000001677f5af500000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000ae7861c80d03826837a50b45aecf11ec677f6586000000000000000000000000000000000000000000000000000000006125b3cc0000000000000000000000000000000000000000000000000000000000000002000000000000000000000000a487bf43cf3b10dffc97a9a744cbb7036965d3b9000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7")));
 
     final contract = tx.getContractInfo();
-    final callInfo = ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+    final callInfo = ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseSwap(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
