@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:typed_data';
 
 import 'package:convert/convert.dart';
 import 'package:decimal/decimal.dart';
@@ -39,8 +40,8 @@ Future<ParsedResult> parseExactInput(ContractCall callInfo, dynamic tx) async {
   }
 
   final args = [
-    [amountIn / Decimal.parse('1e' + tokenIn.params['decimal']), symbolIn],
-    ['≥', amountOutMinimum / Decimal.parse('1e' + tokenOut.params['decimal']), tokenOut.symbol],
+    [(amountIn / Decimal.parse('1e' + tokenIn.params?['decimal'])).toDecimal().toString(), symbolIn],
+    ['≥', (amountOutMinimum / Decimal.parse('1e' + tokenOut.params?['decimal'])).toDecimal().toString(), tokenOut.symbol],
   ];
 
   return ParsedResult(recipient, ['exactInput', args[0], args[1]]);
@@ -69,8 +70,8 @@ Future<ParsedResult> parseExactInputSingle(ContractCall callInfo, dynamic tx) as
   }
 
   final args = [
-    [amountIn / Decimal.parse('1e' + tokenIn.params['decimal']), symbolIn],
-    ['≥', amountOutMinimum / Decimal.parse('1e' + tokenOut.params['decimal']), tokenOut.symbol],
+    [(amountIn / Decimal.parse('1e' + tokenIn.params?['decimal'])).toDecimal().toString(), symbolIn],
+    ['≥', (amountOutMinimum / Decimal.parse('1e' + tokenOut.params?['decimal'])).toDecimal().toString(), tokenOut.symbol],
   ];
 
   return ParsedResult(recipient, ['exactInputSingle', args[0], args[1]]);
@@ -100,8 +101,8 @@ Future<ParsedResult> parseExactOutput(ContractCall callInfo, dynamic tx) async {
   }
 
   final args = [
-    ['≤', amountInMaximum / Decimal.parse('1e' + tokenIn.params['decimal']), symbolIn],
-    [amountOut / Decimal.parse('1e' + tokenOut.params['decimal']), tokenOut.symbol],
+    ['≤', (amountInMaximum / Decimal.parse('1e' + tokenIn.params?['decimal'])).toDecimal().toString(), symbolIn],
+    [(amountOut / Decimal.parse('1e' + tokenOut.params?['decimal'])).toDecimal().toString(), tokenOut.symbol],
   ];
 
   return ParsedResult(recipient, ['exactOutput', args[0], args[1]]);
@@ -130,8 +131,8 @@ Future<ParsedResult> parseExactOutputSingle(ContractCall callInfo, dynamic tx) a
   }
 
   final args = [
-    ['≤', amountInMaximum / Decimal.parse('1e' + tokenIn.params['decimal']), symbolIn],
-    [amountOut / Decimal.parse('1e' + tokenOut.params['decimal']), tokenOut.symbol],
+    ['≤', (amountInMaximum / Decimal.parse('1e' + tokenIn.params?['decimal'])).toDecimal().toString(), symbolIn],
+    [(amountOut / Decimal.parse('1e' + tokenOut.params?['decimal'])).toDecimal().toString(), tokenOut.symbol],
   ];
 
   return ParsedResult(recipient, ['exactOutputSingle', args[0], args[1]]);
@@ -141,7 +142,7 @@ Future<ParsedResult> parseMulticall(ContractCall callInfo, dynamic tx) async {
   final abi = getContractABIByType('UNISWAP V3');
 
   final bytesArray = callInfo.callParams['data'];
-  final subcallArray = bytesArray.map((bytes) => abi.decomposeCall(bytes)).toList();
+  final subcallArray = bytesArray.map((bytes) => abi?.decomposeCall(bytes)).toList();
 
   Set<String> recipients = Set();
   List<dynamic> stack = [];
@@ -194,7 +195,7 @@ Future<ParsedResult> parseMulticall(ContractCall callInfo, dynamic tx) async {
             recipients.add(recipient);
 
             final amountMinimum = Decimal.parse(callInfo.callParams['amountMinimum'].toString());
-            weth[weth.length - 2] = amountMinimum / Decimal.parse('1e18');
+            weth[weth.length - 2] = (amountMinimum / Decimal.parse('1e18')).toDecimal().toString();
           }
         }
         continue;
@@ -244,12 +245,12 @@ void main() async {
         int.parse('40141', radix: 16),
         int.parse('35458b44e', radix: 16),
         int.parse('33', radix: 16),
-        input: hex.decode(
-            "c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000035b3a5601888971afc0b2cadb3630f473332f86b0000000000000000000000000000000000000000000000000000000060ebfda20000000000000000000000000000000000000000000000000000000000004e20000000000000000000000000000000000000000000000000000000015b8aaead0000000000000000000000000000000000000000000000000000000000000042cc8fa225d80b9c7d42f96e9570156c65d6caaa25000bb8c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f4dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000035b3a5601888971afc0b2cadb3630f473332f86b0000000000000000000000000000000000000000000000000000000060ebfda20000000000000000000000000000000000000000000000000000000000004e20000000000000000000000000000000000000000000000000000000015b8aaead0000000000000000000000000000000000000000000000000000000000000042cc8fa225d80b9c7d42f96e9570156c65d6caaa25000bb8c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f4dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseExactInput(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -267,12 +268,12 @@ void main() async {
         int.parse('3fe52', radix: 16),
         int.parse('35458b44e', radix: 16),
         int.parse('33', radix: 16),
-        input: hex.decode(
-            "c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000f12e87fbb4cf2b441ff64a2afe7a027d4e2ed3cf0000000000000000000000000000000000000000000000000000000060efe39200000000000000000000000000000000000000000000000000038d7ea4c680000000000000000000000000000000000000000000000000001b05057b507073860000000000000000000000000000000000000000000000000000000000000042c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f42260fac5e5542a773aa44fbcfedf7c193bc2c599000bb86b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000f12e87fbb4cf2b441ff64a2afe7a027d4e2ed3cf0000000000000000000000000000000000000000000000000000000060efe39200000000000000000000000000000000000000000000000000038d7ea4c680000000000000000000000000000000000000000000000000001b05057b507073860000000000000000000000000000000000000000000000000000000000000042c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f42260fac5e5542a773aa44fbcfedf7c193bc2c599000bb86b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseExactInput(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -290,12 +291,12 @@ void main() async {
         int.parse('2dd3b', radix: 16),
         int.parse('3d432cf59', radix: 16),
         int.parse('3', radix: 16),
-        input: hex.decode(
-            "414bf389000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec700000000000000000000000000000000000000000000000000000000000001f4000000000000000000000000884f4e65b069fc01b2374066f068c2840e9b151d0000000000000000000000000000000000000000000000000000000060ebfe0c000000000000000000000000000000000000000000000000001717b72f0a40000000000000000000000000000000000000000000000000000000000000bdfe8e0000000000000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "414bf389000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec700000000000000000000000000000000000000000000000000000000000001f4000000000000000000000000884f4e65b069fc01b2374066f068c2840e9b151d0000000000000000000000000000000000000000000000000000000060ebfe0c000000000000000000000000000000000000000000000000001717b72f0a40000000000000000000000000000000000000000000000000000000000000bdfe8e0000000000000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseExactInputSingle(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -313,12 +314,12 @@ void main() async {
         int.parse('334da', radix: 16),
         int.parse('324a9a700', radix: 16),
         int.parse('139', radix: 16),
-        input: hex.decode(
-            "f28c0498000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000006f7b3ec82a000c5a310636d7f7549df4256666660000000000000000000000000000000000000000000000000000000060ec0bf8000000000000000000000000000000000000000000000000000000000000038400000000000000000000000000000000000000000000000e869fa41467bc36750000000000000000000000000000000000000000000000000000000000000042cc8fa225d80b9c7d42f96e9570156c65d6caaa25002710c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f46b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "f28c0498000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a00000000000000000000000006f7b3ec82a000c5a310636d7f7549df4256666660000000000000000000000000000000000000000000000000000000060ec0bf8000000000000000000000000000000000000000000000000000000000000038400000000000000000000000000000000000000000000000e869fa41467bc36750000000000000000000000000000000000000000000000000000000000000042cc8fa225d80b9c7d42f96e9570156c65d6caaa25002710c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20001f46b175474e89094c44da98b954eedeac495271d0f000000000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseExactOutput(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -336,12 +337,12 @@ void main() async {
         int.parse('2c679', radix: 16),
         int.parse('2540be400', radix: 16),
         int.parse('1', radix: 16),
-        input: hex.decode(
-            "db3e2198000000000000000000000000b5fe099475d3030dde498c3bb6f3854f762a48ad000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000000000000000000000000000000000000000000bb8000000000000000000000000ba3eb54e93621ec4995fc8a0d8543356b78a01b10000000000000000000000000000000000000000000000000000000060e9511c0000000000000000000000000000000000000000000000000000000002fdfdc00000000000000000000000000000000000000000000000023801a23e5dc9aa4e0000000000000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "db3e2198000000000000000000000000b5fe099475d3030dde498c3bb6f3854f762a48ad000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec70000000000000000000000000000000000000000000000000000000000000bb8000000000000000000000000ba3eb54e93621ec4995fc8a0d8543356b78a01b10000000000000000000000000000000000000000000000000000000060e9511c0000000000000000000000000000000000000000000000000000000002fdfdc00000000000000000000000000000000000000000000000023801a23e5dc9aa4e0000000000000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseExactOutputSingle(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -359,12 +360,12 @@ void main() async {
         int.parse('30cf4', radix: 16),
         int.parse('9502f9000', radix: 16),
         int.parse('25', radix: 16),
-        input: hex.decode(
-            "ac9650d800000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000001800000000000000000000000000000000000000000000000000000000000000104db3e2198000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20000000000000000000000009355372396e3f6daf13359b7b607a3374cc638e00000000000000000000000000000000000000000000000000000000000000bb8000000000000000000000000a18fb9312ef2d98219f169653d8ad9e21df1fabc0000000000000000000000000000000000000000000000000000000060ee624f00000000000000000000000000000000000000000000000000000000000186a000000000000000000000000000000000000000000000000000854de5996b3c58000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000412210e8a00000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "ac9650d800000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000001800000000000000000000000000000000000000000000000000000000000000104db3e2198000000000000000000000000c02aaa39b223fe8d0a0e5c4f27ead9083c756cc20000000000000000000000009355372396e3f6daf13359b7b607a3374cc638e00000000000000000000000000000000000000000000000000000000000000bb8000000000000000000000000a18fb9312ef2d98219f169653d8ad9e21df1fabc0000000000000000000000000000000000000000000000000000000060ee624f00000000000000000000000000000000000000000000000000000000000186a000000000000000000000000000000000000000000000000000854de5996b3c58000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000412210e8a00000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseMulticall(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -382,12 +383,12 @@ void main() async {
         int.parse('32952', radix: 16),
         int.parse('773594000', radix: 16),
         int.parse('fd', radix: 16),
-        input: hex.decode(
-            "ac9650d8000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000000c4f3995c67000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480000000000000000000000000000000000000000000000000000000047a7ed540000000000000000000000000000000000000000000000000000000060efb7ee000000000000000000000000000000000000000000000000000000000000001b7e01e9453c1deaea33180c6ef2dab3659fd844bcf7dcf52bda88bda4c032b8b635273d39a4923755444518a5cf049b2f5b4e4f583d57c24d23fecb7cbf267e76000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000104414bf389000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000c221b7e65ffc80de234bbb6667abdd46593d34f00000000000000000000000000000000000000000000000000000000000000bb80000000000000000000000000d68b16b9c686fe2660c22fa0f01116555147dc70000000000000000000000000000000000000000000000000000000060efb3440000000000000000000000000000000000000000000000000000000047a7ed54000000000000000000000000000000000000000000000054c13c6b81f3a66ef5000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "ac9650d8000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000014000000000000000000000000000000000000000000000000000000000000000c4f3995c67000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480000000000000000000000000000000000000000000000000000000047a7ed540000000000000000000000000000000000000000000000000000000060efb7ee000000000000000000000000000000000000000000000000000000000000001b7e01e9453c1deaea33180c6ef2dab3659fd844bcf7dcf52bda88bda4c032b8b635273d39a4923755444518a5cf049b2f5b4e4f583d57c24d23fecb7cbf267e76000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000104414bf389000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48000000000000000000000000c221b7e65ffc80de234bbb6667abdd46593d34f00000000000000000000000000000000000000000000000000000000000000bb80000000000000000000000000d68b16b9c686fe2660c22fa0f01116555147dc70000000000000000000000000000000000000000000000000000000060efb3440000000000000000000000000000000000000000000000000000000047a7ed54000000000000000000000000000000000000000000000054c13c6b81f3a66ef5000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseMulticall(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
@@ -405,12 +406,12 @@ void main() async {
         int.parse('54206', radix: 16),
         int.parse('df8475800', radix: 16),
         int.parse('90b', radix: 16),
-        input: hex.decode(
-            "ac9650d8000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000002e000000000000000000000000000000000000000000000000000000000000000c4f3995c67000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480000000000000000000000000000000000000000000000000000000084e57a320000000000000000000000000000000000000000000000000000000060ee6b26000000000000000000000000000000000000000000000000000000000000001cc6b0986f2aefb2705902e9a570a6a381984987184dc961608204eba25dba2bd23a25f0d4234f7ac6512175dd73060f93d0f2d0fe3b6e03f557f183fc3aed904c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000144c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000060ee66760000000000000000000000000000000000000000000000000000000084e57a32000000000000000000000000000000000000000000000000105dd2eee13177830000000000000000000000000000000000000000000000000000000000000042a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480001f46b175474e89094c44da98b954eedeac495271d0f0001f4c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004449404b7c000000000000000000000000000000000000000000000000105dd2eee1317783000000000000000000000000916c2cd2f87c0e9d7a02b76aa8ba6cb6586187d400000000000000000000000000000000000000000000000000000000"));
+        input: Uint8List.fromList(hex.decode(
+            "ac9650d8000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000030000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000002e000000000000000000000000000000000000000000000000000000000000000c4f3995c67000000000000000000000000a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480000000000000000000000000000000000000000000000000000000084e57a320000000000000000000000000000000000000000000000000000000060ee6b26000000000000000000000000000000000000000000000000000000000000001cc6b0986f2aefb2705902e9a570a6a381984987184dc961608204eba25dba2bd23a25f0d4234f7ac6512175dd73060f93d0f2d0fe3b6e03f557f183fc3aed904c000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000144c04b8d59000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000a000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000060ee66760000000000000000000000000000000000000000000000000000000084e57a32000000000000000000000000000000000000000000000000105dd2eee13177830000000000000000000000000000000000000000000000000000000000000042a0b86991c6218b36c1d19d4a2e9eb0ce3606eb480001f46b175474e89094c44da98b954eedeac495271d0f0001f4c02aaa39b223fe8d0a0e5c4f27ead9083c756cc200000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004449404b7c000000000000000000000000000000000000000000000000105dd2eee1317783000000000000000000000000916c2cd2f87c0e9d7a02b76aa8ba6cb6586187d400000000000000000000000000000000000000000000000000000000")));
 
     final contract = tx.getContractInfo();
     final callInfo =
-        ContractCall.fromJson({'function': contract['method'], 'params': contract['params']});
+        ContractCall.fromJson({'function': contract?['method'], 'params': contract?['params']});
 
     final parsed = await parseMulticall(callInfo, tx);
     expect(parsed.recipient, tx.from.toString());
